backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
google-genai>=1.0.0
faiss-cpu==1.13.2
numpy==1.26.4
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==10.4.0
yt-dlp==2024.8.6
youtube-transcript-api==0.6.2
pydub==0.25.1
SpeechRecognition==3.10.4
aiofiles==24.1.0
pydantic==2.9.1
pydantic-settings==2.5.2
httpx==0.27.2
beautifulsoup4==4.12.3
